feat: add layer shell
This uses wlroots layer shell to create a surface. This surface is bound
to the top of the screen (this means left, top, and right). The surface
is currently roughly the size of a bar that is placed on the top of a
2560x1440 screen.

The surface has an exclusive zone, so other windows will be pushed out
of its way.